% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popmean_function.R
\name{popmean}
\alias{popmean}
\title{Title calculates the national prevalence by year and by intervention}
\usage{
popmean(data, intervention = "IRS", years = 2020:2022)
}
\arguments{
\item{data}{a dataset with columns "zone", "population", "year", "none", "IRS", "ITN", "IRS.ITN"}

\item{intervention}{intervention type. It must be one of the following intervention:  "none", "IRS", "ITN", "IRS.ITN".}

\item{years}{a numerical vector. It must contain years for wich the national prevalence have to be calculate.}
}
\description{
Title calculates the national prevalence by year and by intervention
}
